Move helper-code to helper.py
